LPS-180472 Rename files
