+++ modules/apps/segments/segments-web/src/main/resources/META-INF/resources/js/SegmentsApp.es.js
@@ -15,7 +15,7 @@
 import React from 'react';
 
 import ThemeContext from './ThemeContext.es';
-import SegmentEdit from './components/segment_edit/SegmentEdit.es';
+import SegmentEdit from './components/segment_edit/SegmentEdit';
 
 export default function ({context, props}) {
 	return (
