+++ modules/apps/segments/segments-web/test/js/components/criteria_sidebar/CriteriaSidebar.test.js
@@ -15,7 +15,7 @@
 import {cleanup, render} from '@testing-library/react';
 import React from 'react';
 
-import CriteriaSidebar from '../../../../src/main/resources/META-INF/resources/js/components/criteria_sidebar/CriteriaSidebar.es';
+import CriteriaSidebar from '../../../../src/main/resources/META-INF/resources/js/components/criteria_sidebar/CriteriaSidebar';
 
 describe('CriteriaSidebar', () => {
 	afterEach(cleanup);
