+++ modules/apps/segments/segments-web/test/js/components/criteria_sidebar/CriteriaSidebarItem.test.js
@@ -15,7 +15,7 @@
 import {cleanup, render} from '@testing-library/react';
 import React from 'react';
 
-import CriteriaSidebarItem from '../../../../src/main/resources/META-INF/resources/js/components/criteria_sidebar/CriteriaSidebarItem.es';
+import CriteriaSidebarItem from '../../../../src/main/resources/META-INF/resources/js/components/criteria_sidebar/CriteriaSidebarItem';
 
 const connectDnd = jest.fn((element) => element);
